Adding a check to not show the "Page #" title when there is only one.
Motivation:
Currently when showing a forms app that only has one page, and does not
have a custom pageName, a title for the view is shown as 'Page 1'. When
there is only one page then we should not show this page name at all.

Modifications:
Added checks to see if there is a pageName set but

Result:
All test pass:
$ cd src/appforms
$ grunt testServer
Open http://127.0.0.1:3001/?grep=Backbone%20-%20Step%20View

JIRA:
https://issues.jboss.org/browse/RHMAP-4762