Removing all empty backbone tests.
Motivation:
Currently there are quite a few tests in
src/appforms/tests/tests/backbone that are empty which might cause some
confusion.
There are a few that are not empty and contain test code but that are
not included in modues.js hence they are not being run by testServer.

Modifications:
$ find . -type f -empty | xargs git rm

Result:
Project is a little cleaner.