Recreate mock_uuid cookie if blank string
Currently when the value of a cookie is blank it is read as an empty
string. This passes the current check that is looking for the value
being `null`. If the value is `null` the cookie will be recreated.

This behaviour will cause fh-mbaas-api to throw an error, as it checks
whether the CUID sent from the client is a falsy value.

This changes the check on the client SDK to match that of fh-mbaas-api
so that empty-string device ids will be recreated as well as `null` ones.